rankify/requirements.txt
pandas==2.2.3
prettytable==3.11.0
tqdm==4.66.5
requests==2.32.3
torch==2.5.0
transformers==4.45.2
sentencepiece==0.2.0
openai==1.52.2
anthropic==0.37.1
together==1.3.3
onnxruntime==1.19.2
llama-cpp-python==0.2.76
vllm==0.6.3
ftfy==6.3.1
dacite==1.8.1
fschat[model_worker]>=0.2.36
llm-blender==0.0.2
sentence_transformers==3.3.0
flash-attn==2.5.0
pyserini==0.44.0
faiss-cpu==1.9.0.post1
omegaconf==2.3.0
h5py==3.12.1
py7zr==0.22.0
ujson==5.10.0
ninja==1.11.1.3
cohere==5.14.0
dotenv>=0.9.9
litellm>=1.61.20
langchain>=0.3.19
fasttext-wheel>=0.9.2
wikipedia-api>=0.8.1
pillow>=10.4.0
smolagents>=1.9.2
Crawl4AI>=0.6.3